# Use span copy in BigInteger.TryGetBytes when applicable