When the value is positive (we don't need to do two's complement conversion) and on a little endian machine, we can copy the bytes out from the underlying uint[] storage.

| Method      | Job        | BitLength | IsBigEndian | Mean     | Error    | StdDev   | Ratio |
|------------ |----------- |---------- |------------ |---------:|---------:|---------:|------:|
| **ToByteArray** | **PR** | **1024**      | **False**       | **21.64 ns** | **0.020 ns** | **0.015 ns** |  **0.41** |
| ToByteArray | main | 1024      | False       | 53.38 ns | 0.157 ns | 0.140 ns |  1.00 |
|             |            |           |             |          |          |          |       |
| **ToByteArray** | **PR** | **1024**      | **True**        | **23.45 ns** | **0.021 ns** | **0.018 ns** |  **0.44** |
| ToByteArray | main | 1024      | True        | 53.48 ns | 0.150 ns | 0.125 ns |  1.00 |

<details>

```c#
public class Benchmarks
{
    [Params(1024)]
    public int BitLength;

    [Params(true, false)]
    public bool IsBigEndian;

    [Params(true)]
    public bool IsPositive;

    private BigInteger n;

    [GlobalSetup]
    public void Setup()
    {
        Random r = new(123);
        byte[] bytes = new byte[BitLength / 8];
        r.NextBytes(bytes);
        if (IsPositive)
        {
            bytes[^1] &= 0x7F;
        }
        else
        {
            bytes[^1] |= 0x80;
        }
        n = new BigInteger(bytes);
    }

    [Benchmark] public byte[] ToByteArray() => n.ToByteArray(isBigEndian: IsBigEndian);
}
```

</details>