@Rob-Hague 2025-05-10
@huoyaoyuan `isBigEndian` is a parameter passed to [ToByteArray](https://learn.microsoft.com/dotnet/api/system.numerics.biginteger.tobytearray?view=net-9.0#system-numerics-biginteger-tobytearray(system-boolean-system-boolean))